api/requirement.txt
Flask
torch
transformers
Pillow

